Letting people know how to embed this to Meteor
I don't know much about sass/scss but I was looking for a way to embed bidirectional Bootstrap to Meteor, or any other framework which has a bidirectional functionality for RTL/LTR support.
I tried Foundation from Zurb, but for me, as I guess for most of ppl like me using Meteor, it's all a headache going through the learning curve of Foundation Zurb, and also learning how to embed it correctly in to meteor, since embedding it means to have call js onCreated and onDestroyed for each template instance to create certain components out of this framework to really work in meteor. $(document).foundation() simply wont work ever.
All these headaches for a single functionality of Bidirectional RTL/LTR Support was not true.
Ending up at your repo, I gave it a try, and found the way to use it in Meteor, I already have it in action with a project I'm working on now.
I'll do create a sample git repo if I could in future.
I did the things above, but since you know scss I was wondering if you could shorten the `main.scss` file content as described above, to something one-liner. Please keep in mind and I found this structure (having _bidirectional.scss on top) worked.
I'd be happy if you work on it.
Thanks for your great work on the project too.
Waiting for you next release on v4